Inherit map from base class
