+++ angular-app/src/app/components/base-map/base-map.component.css
@@ -1,0 +1,7 @@
+.full-height {
+    min-height: 100%;
+}
+
+.locator-icon {
+    position: absolute;
+}
